🔨 Change to `master` branch
This blog started when Github pages was very different than it is today.
It used to only be able to deploy from a branch named `gh-pages` or from
a `docs/` folder; and so this project opted for an esoteric setup
using a `source` branch and a `master` branch to enable this.

However, with the years since the start of this blog, we can now deploy
with Github artifacts -- something that wasn't even concieved at the time.
As a result, this branch is now named `master`, like the other repos.
Yay!